Implemented floating point drawing for 2D plot objetcs.
git-svn-id: http://svn.berlios.de/svnroot/repos/qtiplot/trunk@827 5a6a7de5-fb12-0410-b871-c33778c25c60

+++ qtiplot/src/VectorCurve.h
@@ -64,8 +64,8 @@ public:
 	void setVectorEnd(const QString& xColName, const QString& yColName);
 	void setVectorEnd(const QwtArray<double>&x, const QwtArray<double>&y);
 
-	int width();
-	void setWidth(int w);
+	double width();
+	void setWidth(double w);
 
 	QColor color();
 	void setColor(const QColor& c);
